--- pkg/proofs/sigma/compilerutils/utils.go
@@ -1,25 +1,0 @@
-package compilerutils
-
-import (
-	"io"
-
-	"github.com/bronlabs/bron-crypto/pkg/base/errs"
-	"github.com/bronlabs/bron-crypto/pkg/proofs/sigma"
-	"github.com/bronlabs/bron-crypto/pkg/proofs/sigma/compiler"
-	"github.com/bronlabs/bron-crypto/pkg/proofs/sigma/compiler/fiatshamir"
-	"github.com/bronlabs/bron-crypto/pkg/proofs/sigma/compiler/fischlin"
-	"github.com/bronlabs/bron-crypto/pkg/proofs/sigma/compiler/randfischlin"
-)
-
-func Compile[X sigma.Statement, W sigma.Witness, A sigma.Statement, S sigma.State, Z sigma.Response](compilerName compiler.Name, sigmaProtocol sigma.Protocol[X, W, A, S, Z], prng io.Reader) (compiler.NICompiler[X, W], error) {
-	switch compilerName {
-	case fiatshamir.Name:
-		return fiatshamir.NewCompiler(sigmaProtocol)
-	case fischlin.Name:
-		return fischlin.NewCompiler(sigmaProtocol, prng)
-	case randfischlin.Name:
-		return randfischlin.NewCompiler(sigmaProtocol, prng)
-	default:
-		return nil, errs.NewArgument("unknown compiler name: %s", compilerName)
-	}
-}
